+++ NEWS.md
@@ -40,7 +40,7 @@
 Next steps: 
 - [x] `Validation: VALID` is super vague and I don't think there's acyclicity-check
     functionality in `validate.R`. Fix that output and add acyclicity function.
-- [ ] Adapt `print.DAGassist_report` to work with different models like feols, 
+- [x] Adapt `print.DAGassist_report` to work with different models like feols, 
     specifically in the model compare. feols is snagging on "original fit 
     (coef head)". might be time to break out `stargazer` or `modelsummary`.
 - [ ] Integrate canonical models